Moved asset types initialisation method into the business layer
