+++ source/app/business/asset_types.py
@@ -1,0 +1,28 @@
+#  IRIS Source Code
+#  Copyright (C) ${current_year} - DFIR-IRIS
+#  contact@dfir-iris.org
+#
+#  This program is free software; you can redistribute it and/or
+#  modify it under the terms of the GNU Lesser General Public
+#  License as published by the Free Software Foundation; either
+#  version 3 of the License, or (at your option) any later version.
+#
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+#  Lesser General Public License for more details.
+#
+#  You should have received a copy of the GNU Lesser General Public License
+#  along with this program; if not, write to the Free Software Foundation,
+#  Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+from app.models.models import AssetsType
+
+
+def create_asset_type_if_not_exists(session, **kwargs):
+    instance = session.query(AssetsType).filter_by(asset_name=kwargs.get('asset_name')).first()
+    if instance:
+        return
+
+    instance = AssetsType(**kwargs)
+    session.add(instance)
+    session.commit()
